+++ packages/terra-list/src/ListSubsection.jsx
@@ -61,6 +61,10 @@ const propTypes = {
    * The intl object to be injected for translations.
    */
   intl: PropTypes.shape({ formatMessage: PropTypes.func }),
+  /**
+   * z-index value for the list item (li element). Defaults to 6001 which is greater value than terra-modal-manager z-index value.
+   */
+  zIndex: PropTypes.number,
 };
 
 const defaultProps = {
@@ -68,6 +72,7 @@ const defaultProps = {
   isCollapsed: false,
   isCollapsible: false,
   level: 2,
+  zIndex: 6001,
 };
 
 const ListSubsection = ({
@@ -77,12 +82,17 @@ const ListSubsection = ({
   isDraggable,
   onDragEnd,
   intl,
+  zIndex,
   ...customProps
 }) => {
   const [listItemNodes, setlistItemNodes] = useState(children);
   let listSubSectionItemNode = useRef();
+  const isListItemDropped = useRef();
+  const draggedItemindex = useRef();
+
   /* eslint-disable-next-line no-param-reassign */
   delete customProps?.isTabFocusDisabled;
+
   useEffect(() => {
     if (!isCollapsible || !isCollapsed) {
       if (Array.isArray(children)) {
@@ -94,6 +104,16 @@ const ListSubsection = ({
       setlistItemNodes([]);
     }
   }, [children, isCollapsible, isCollapsed]);
+
+  useEffect(() => {
+    if (isListItemDropped.current) {
+      const listItems = listSubSectionItemNode && listSubSectionItemNode.querySelectorAll('[data-item-show-focus]');
+      if (listItems[draggedItemindex.current]) {
+        listItems[draggedItemindex.current].focus();
+      }
+      isListItemDropped.current = false;
+    }
+  }, [listItemNodes]);
 
   const theme = useContext(ThemeContext);
   const listClassNames = classNames(
@@ -155,12 +175,7 @@ const ListSubsection = ({
       result.destination.index,
     );
     setlistItemNodes(items);
-    if (listSubSectionItemNode) {
-      const listitems = listSubSectionItemNode.querySelectorAll('[data-item-show-focus]');
-      if (listitems && listitems[result.source.index]) {
-        listitems[result.source.index].focus();
-      }
-    }
+    draggedItemindex.current = result.destination.index;
     provided.announce(intl.formatMessage({ id: 'Terra.list.drop' }, { startPosition: (result.source.index + 1), endPosition: (result.destination.index + 1) }));
     if (onDragEnd) {
       onDragEnd(result, provided);
@@ -168,6 +183,7 @@ const ListSubsection = ({
   };
 
   const handleDragStart = (start, provided) => {
+    isListItemDropped.current = true;
     provided.announce(intl.formatMessage({ id: 'Terra.list.lift' }, { startPosition: (start.source.index + 1) }));
   };
 
@@ -177,13 +193,22 @@ const ListSubsection = ({
     }
   };
 
-  const cloneListItem = (ListItem, provider) => React.cloneElement(ListItem, {
+  const getStyleforDrag = (ListItem, snapshot, provider) => {
+    const styleProperties = provider.draggableProps.style;
+    if (styleProperties && snapshot && snapshot.isDragging) {
+      styleProperties.zIndex = zIndex;
+    }
+    return styleProperties;
+  };
+
+  const cloneListItem = (ListItem, provider, snapshot) => React.cloneElement(ListItem, {
     isDraggable: ListItem?.props?.isSelectable,
     refCallback: (refobj) => {
       provider.innerRef(refobj);
     },
     ...provider.draggableProps,
     ...provider.dragHandleProps,
+    style: getStyleforDrag(ListItem, snapshot, provider),
   });
 
   const renderSubSectionListItemsDom = () => (
@@ -200,12 +225,14 @@ const ListSubsection = ({
     </>
   );
 
+  window['__react-beautiful-dnd-disable-dev-warnings'] = true;
+
   const renderDraggableListDom = () => (
     <DragDropContext onDragEnd={handleDragEnd} onDragStart={handleDragStart} onDragUpdate={handleDragUpdate}>
       <Droppable
         droppableId="listSubSection"
         renderClone={(provided, snapshot, rubric) => (
-          cloneListItem(listItemNodes[rubric.source.index], provided)
+          cloneListItem(listItemNodes[rubric.source.index], provided, snapshot)
         )}
       >
         {(provided) => (
@@ -220,8 +247,8 @@ const ListSubsection = ({
               >
                 {listItemNodes.map((item, index) => (
                   <Draggable isDragDisabled={!(item?.props?.isSelectable)} key={item.key} draggableId={item.key} index={index}>
-                    {(provider) => (
-                      cloneListItem(item, provider)
+                    {(provider, snapshot) => (
+                      cloneListItem(item, provider, snapshot)
                     )}
                   </Draggable>
                 ))}